+++ tests/Structures/VehicleModelTest.php
@@ -20,16 +20,16 @@ class VehicleModelTest extends AbstractStructureTestCase
     public function testConfigure(): void
     {
         $this->instance = $this->factory($input = [
-            'id'      => $description = 'some description',
             'name'    => $name = 'some name',
+            'id'      => $description = 'some description',
             'mark_id' => $mark_id = 'some mark id',
         ]);
 
-        $this->assertEquals($description, $this->instance->getId());
-        $this->assertEquals($name, $this->instance->getName());
-        $this->assertEquals($mark_id, $this->instance->getMarkId());
+        $this->assertSame($description, $this->instance->getId());
+        $this->assertSame($name, $this->instance->getName());
+        $this->assertSame($mark_id, $this->instance->getMarkId());
 
-        $this->assertEquals($input, $this->instance->toArray());
+        $this->assertSame($input, $this->instance->toArray());
     }
 
     /**
@@ -58,7 +58,7 @@ class VehicleModelTest extends AbstractStructureTestCase
         ]);
 
         $this->assertTrue(isset($this->instance['name']));
-        $this->assertEquals($name, $this->instance['name']);
+        $this->assertSame($name, $this->instance['name']);
 
         $this->assertFalse(isset($this->instance['bar']));
     }